+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.11.0] - 2021-10-21
+
+### Changed
+
+- Made animation system intern to the spatial object class
+- Made the animation manager affect all animations
+
+### Removed
+
+- Removed Bob
+
 ## [0.9.0] - 2021-10-13
 
 ### Added